remove pytest import
